add: Architectury and project modules
- Use Architectury
- Add Forge mod jar
- Minor changes
