+++ .architectury-transformer/debug.log
@@ -1,0 +1,1 @@
+[Architectury Transformer DEBUG] Closing File Systems for D:\Coding\My Projects\FabricAntiXray\common\build\libs\anti-xray-1.2.0.jar
