+++ .github/workflows/build.yml
@@ -20,5 +20,6 @@ jobs:
       - name: capture build artifacts
         uses: actions/upload-artifact@v2
         with:
-          name: artifacts
-          path: build/libs/
+          name: Build-Artifacts
+          path: |
+            **/build/libs/
