+++ build.gradle
@@ -1,94 +1,45 @@
 plugins {
-    id 'fabric-loom' version '0.10-SNAPSHOT'
-    id 'maven-publish'
+    id "architectury-plugin" version "3.4-SNAPSHOT" // Uses Architectury plugin https://github.com/architectury/architectury-plugin
+    id "dev.architectury.loom" version "0.10.0-SNAPSHOT" apply false // Uses Architectury Loom
+    // (fork of Fabric Loom) https://github.com/architectury/architectury-loom - does not apply to root project
 }
 
-sourceCompatibility = JavaVersion.VERSION_17
-targetCompatibility = JavaVersion.VERSION_17
-
-archivesBaseName = "${project.archives_base_name}-mc${project.minecraft_version}"
-version = project.mod_version
-group = project.maven_group
-
-repositories {
-    // Add repositories to retrieve artifacts from in here.
-    // You should only use this when depending on other mods because
-    // Loom adds the essential maven repositories to download Minecraft and libraries from automatically.
-    // See https://docs.gradle.org/current/userguide/declaring_repositories.html
-    // for more information about repositories.
+architectury {
+    minecraft = rootProject.minecraft_version // Sets Minecraft version from gradle.properties
 }
 
-loom {
-    accessWidenerPath = file("src/main/resources/antixray.accesswidener")
-}
+subprojects {
+    apply plugin: "dev.architectury.loom" // Applies Architectury Loom to subprojects
 
-dependencies {
-    minecraft "com.mojang:minecraft:${project.minecraft_version}"
-    modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
-    mappings minecraft.officialMojangMappings()
-
-    // Fabric API
-    Set<String> apiModules = [
-            'fabric-api-base',
-            'fabric-lifecycle-events-v1'
-    ]
-    apiModules.forEach {
-        modImplementation(fabricApi.module(it, project.fabric_version))
+    loom {
+        silentMojangMappingsLicense() // Silences the annoying as hell Mojang License text
     }
 
-    // Toml
-    implementation(include("com.moandjiezana.toml:toml4j:${project.toml_version}"))
-}
-
-processResources {
-    inputs.property "version", project.version
-    filesMatching("fabric.mod.json") {
-        expand "version": project.version
+    dependencies {
+        minecraft "com.mojang:minecraft:${rootProject.minecraft_version}" // Minecraft dependency
+        mappings loom.officialMojangMappings() // Maps Minecraft so we can tell what classes are what - official mappings
     }
 }
 
-tasks.withType(JavaCompile).configureEach {
-    // ensure that the encoding is set to UTF-8, no matter what the system default is
-    // this fixes some edge cases with special characters not displaying correctly
-    // see http://yodaconditions.net/blog/fix-for-java-file-encoding-problems-with-gradle.html
-    // If Javadoc is generated, this must be specified in that task too.
-    it.options.encoding = "UTF-8"
+allprojects {
+    apply plugin: "java"
+    apply plugin: "architectury-plugin" // Applies architectury plugin to all projects
 
-    it.options.release = 17
-}
+    archivesBaseName = rootProject.archives_base_name // sets the archive name from gradle.properties
+    version = rootProject.mod_version // sets the version from gradle.properties
+    group = rootProject.maven_group // sets the maven group from gradle.properties
 
-java {
-    // Loom will automatically attach sourcesJar to a RemapSourcesJar task and to the "build" task
-    // if it is present.
-    // If you remove this line, sources will not be generated.
-    withSourcesJar()
-}
-
-jar {
-    from("LICENSE") {
-        rename { "${it}_${project.archivesBaseName}" }
-    }
-}
-
-// configure the maven publication
-publishing {
-    publications {
-        mavenJava(MavenPublication) {
-            // add all the jars that should be included when publishing to maven
-            artifact(remapJar) {
-                builtBy remapJar
-            }
-            artifact(sourcesJar) {
-                builtBy remapSourcesJar
-            }
-        }
+    java {
+        withSourcesJar()
     }
 
-    // See https://docs.gradle.org/current/userguide/publishing_maven.html for information on how to set up publishing.
-    repositories {
-        // Add repositories to publish to here.
-        // Notice: This block does NOT have the same function as the block in the top level.
-        // The repositories here will be used for publishing your artifact, not for
-        // retrieving dependencies.
+    tasks.withType(JavaCompile).configureEach {
+        it.options.encoding = "UTF-8"
+        it.options.release = 17
     }
+
+    compileJava {
+        options.compilerArgs +=	"-Xlint:-processing" // Fixes "no processor claimed any of these annotations" warning
+    }
+
 }
