+++ common/src/main/java/me/drex/antixray/interfaces/IPalettedContainer.java
