+++ common/src/main/java/me/drex/antixray/util/BitStorageReader.java
