+++ common/src/main/java/me/drex/antixray/util/BitStorageWriter.java
