+++ common/src/main/java/me/drex/antixray/util/Platform.java
@@ -1,0 +1,5 @@
+package me.drex.antixray.util;
+
+public enum Platform {
+    FABRIC, FORGE;
+}
