+++ fabric/src/main/resources/antixray.accesswidener
