+++ forge/src/main/resources/META-INF/accesstransformer.cfg
@@ -1,0 +1,14 @@
+# antixray
+public net.minecraft.world.level.chunk.PaletteResize
+public net.minecraft.world.level.chunk.PalettedContainer$Data
+public net.minecraft.world.level.chunk.PalettedContainer$Configuration
+
+public net.minecraft.server.level.ServerChunkCache m_8364_(J)Lnet/minecraft/server/level/ChunkHolder; # getVisibleChunkIfPresent
+public net.minecraft.world.level.chunk.PalettedContainer$Strategy m_188151_(Lnet/minecraft/core/IdMap;I)I # calculateBitsForSerialization
+
+public net.minecraft.world.level.chunk.PalettedContainer$Strategy f_188139_ # GLOBAL_PALETTE_FACTORY
+public net.minecraft.world.level.chunk.ChunkAccess f_187611_ # levelHeightAccessor
+public net.minecraft.server.level.ServerPlayerGameMode f_9244_ # level
+
+# networking
+public net.minecraft.network.Connection$PacketHolder
