+++ gradle.properties
@@ -1,13 +1,13 @@
-# Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
-# Fabric Properties
-# check these on https://fabricmc.net/versions.html
+org.gradle.jvmargs=-Xmx2048M
+
 minecraft_version=1.18
-loader_version=0.12.8
-# Mod Properties
-mod_version=1.1.2
+
+# Mod properties
+mod_version=1.2.0
 maven_group=me.drex
 archives_base_name=anti-xray
+
 # Dependencies
-fabric_version=0.43.1+1.18
+forge_version=1.18-38.0.17
+fabric_loader_version=0.12.12
 toml_version=0.7.2
