--- src/main/java/me/drex/antixray/AntiXray.java
@@ -1,25 +1,0 @@
-package me.drex.antixray;
-
-import me.drex.antixray.config.Config;
-import net.fabricmc.api.DedicatedServerModInitializer;
-import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
-import net.fabricmc.loader.api.FabricLoader;
-import net.minecraft.server.MinecraftServer;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public class AntiXray implements DedicatedServerModInitializer {
-
-    public static final Logger LOGGER = LogManager.getLogger();
-    private static MinecraftServer minecraftServer;
-
-    public static MinecraftServer getMinecraftServer() {
-        return minecraftServer;
-    }
-
-    @Override
-    public void onInitializeServer() {
-        ServerLifecycleEvents.SERVER_STARTING.register(server -> minecraftServer = server);
-        Config.loadConfig(FabricLoader.getInstance().getConfigDir().resolve("antixray.toml").toFile());
-    }
-}
