--- src/main/resources/networking.mixins.json
@@ -1,13 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "me.drex.networking.mixin",
-  "compatibilityLevel": "JAVA_16",
-  "server": [
-    "ConnectionMixin",
-    "ConnectionPacketHolderAccessor"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
